+++ srcts/types/src/shiny/shinyapp.d.ts
@@ -51,6 +51,7 @@ declare class ShinyApp {
     };
     $nextRequestId: number;
     $allowReconnect: boolean | "force";
+    $hardDisconnectMessage: string | null;
     constructor();
     connect(initialInput: InputValues): void;
     isConnected(): boolean;
@@ -66,6 +67,7 @@ declare class ShinyApp {
         next: () => number;
         reset: () => void;
     };
+    $enterClosedState(): void;
     onDisconnected(reloading?: boolean): void;
     onConnected(): void;
     makeRequest(method: string, args: unknown[], onSuccess: OnSuccessRequest, onError: OnErrorRequest, blobs: Array<ArrayBuffer | Blob | string> | undefined): void;